+++ wasm/CHANGELOG.md
@@ -1,9 +1,11 @@
 Latest online documentation available at: https://kaspa.aspectron.org/docs/
 
-### Latest
+### Release 2024-03-31
 
 - Rename `kaspa-beacon` app to `kaspa-resolver`
-- Change UtxoProcessor event handlers in typescript to receive typed event data
+- Change RpcClient, UtxoProcessor and Wallet event handlers in typescript to receive typed event data
+- UtxoProcessor and Wallet event handlers now deliver TransactionRecord events (Discovery, Pending, etc.)
+as Rust or WASM objects, allowing user to call `hasAddress(<address>)` on the received `event.data.record` object.
 
 ### Release 2024-03-19